+++ src/Executor/Executor.php
@@ -80,51 +80,43 @@ abstract class Executor implements EventListenerInterface
      *
      * For example, with `mariadb` it should return `/usr/bin/mariadb`.
      *
-     * It first checks and returns any value set by the configuration. If not present, it uses `ExecutableFinder::find)`.
+     * It first checks and returns any value set by the configuration. If not present, it uses `ExecutableFinder::find()`.
      * If the binary cannot be found, an exception is thrown.
-     *
-     * You can specify more than one name (for example, if there are possible aliases or fallbacks). In this case, the
-     *  first one found is returned.
      *
      * To use `findBinary()` in conjunction with `getBinaryName()`:
      * ```
-     * $this->findBinary(...(array)$this->getBinaryName())
+     * $this->findBinary($this->getBinaryName())
      * ```
      *
-     * @param \DatabaseBackup\Compression|string ...$name
+     * @param \DatabaseBackup\Compression|string $binaryName
      * @return string
-     * @since 3.0.0
      * @throws \InvalidArgumentException
+     * @since 3.0.0
      */
-    public function findBinary(Compression|string ...$name): string
+    public function findBinary(Compression|string $binaryName): string
     {
-        // Makes sure it doesn't contain `Compression::None`
-        if (array_any(array: $name, callback: fn (Compression|string $name): bool => $name instanceof Compression && !$name->isValid())) {
-            throw new InvalidArgumentException('Unable to search for binary for "none" Compression');
+        if ($binaryName instanceof Compression) {
+            if (!$binaryName->isValid()) {
+                //If it is a `Compression`, it checks if it is a valid `Compression`
+                throw new InvalidArgumentException('Unable to search for binary for "none" Compression');
+            }
+
+            $binaryName = lcfirst($binaryName->name);
         }
 
-        $name = array_map(
-            callback: fn (Compression|string $name): string => $name instanceof Compression ? lcfirst($name->name) : $name,
-            array: $name
+        $binary = Configure::read(
+            var: 'DatabaseBackup.binaries.' . $binaryName,
+            default: new ExecutableFinder()->find(name: $binaryName)
         );
-
-        $ExecutableFinder = new ExecutableFinder();
-
-        foreach ($name as $sName) {
-            $binary = Configure::read(
-                var: 'DatabaseBackup.binaries.' . $sName,
-                default: $ExecutableFinder->find(name: $sName)
-            );
-            if ($binary) {
-                return $binary;
-            }
+        if ($binary) {
+            return $binary;
         }
 
         throw new InvalidArgumentException(__d(
             'database_backup',
             'Binary for `{0}` could not be found. You have to set its path manually on your bootstrap with: `{1}`',
-            $name[0],
-            'Configure::write(\'DatabaseBackup.binaries.' . $name[0] . '\', \'/your/full/path/to/' . $name[0] . '\')'
+            $binaryName,
+            'Configure::write(\'DatabaseBackup.binaries.' . $binaryName . '\', \'/your/full/path/to/' . $binaryName . '\')'
         ));
     }